+++ src/Socket/Listen.cpp
@@ -6,13 +6,13 @@
 /*   By: agimi <agimi@student.1337.ma>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/03 10:35:32 by agimi             #+#    #+#             */
-/*   Updated: 2024/02/14 12:35:03 by agimi            ###   ########.fr       */
+/*   Updated: 2024/02/14 15:04:58 by agimi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <webserv.hpp>
 
-wbs::Listen::Listen(int d, int s, int pro, int por, u_long i, int b) : Bind(d, s, pro, por, i), blog(b)
+wbs::Listen::Listen(int d, int s, int pro, int por, unsigned int i, int b) : Bind(d, s, pro, por, i), blog(b)
 {
 	listning();
 	c_test(list);
